+++ packages/react/src/components/MenuButton/src/menuButton.stories.tsx
@@ -1,15 +1,16 @@
-import React, { forwardRef } from "react";
+import React, { useState } from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { useArgs } from "@storybook/preview-api";
 
 import { ChevronRightIcon, HomeIcon } from "@nimbus-ds/icons";
-import { Tag, Badge, Icon } from "@nimbus-ds/components";
+import { Tag, Badge, Icon, Box } from "@nimbus-ds/components";
 
+import { MenuExpandContext } from "@common/contexts";
 import { MenuButton, MenuButtonProps } from "./MenuButton";
 
-export const Basic: React.FC<MenuButtonProps> = forwardRef(
-  ({ ...props }: MenuButtonProps) => <MenuButton {...props} />
-) as React.FC<MenuButtonProps>;
+export const Basic: React.FC<MenuButtonProps> = (props: MenuButtonProps) => (
+  <MenuButton {...props} />
+);
 Basic.displayName = "MenuButton";
 
 const meta: Meta<typeof Basic> = {
@@ -45,9 +46,10 @@ export const sampleAppMenuButton: Story = {
 
 export const stressedAppMenuButton: Story = {
   args: {
-    label: "This is a very stressed menu button with a very long text to showcase what happens when the text overflows the button",
+    label:
+      "This is a very stressed menu button with a very long text to showcase what happens when the text overflows the button",
     startIcon: HomeIcon,
-    children: <Tag appearance="primary">Children tag</Tag>
+    children: <Tag appearance="primary">Children tag</Tag>,
   },
 };
 
@@ -92,5 +94,52 @@ export const menuButtonWithRightIcon: Story = {
     label: "Button with right icon children",
     startIcon: HomeIcon,
     children: <Icon source={<ChevronRightIcon />} />,
+  },
+};
+
+export const menuButtonCollapsed: Story = {
+  render: (args) => {
+    const [expanded, setExpanded] = useState(false);
+    return (
+      <MenuExpandContext.Provider value={{ expanded }}>
+        Click the button to {expanded ? "collapse" : "expand"}
+        <Box maxWidth="200px">
+          <MenuButton {...args} onClick={() => setExpanded(!expanded)} />
+        </Box>
+      </MenuExpandContext.Provider>
+    );
+  },
+  args: {
+    label: "Menu button",
+    startIcon: HomeIcon,
+  },
+};
+
+export const accordionMenuButtonCollapsed: Story = {
+  render: (args) => {
+    const [expanded, setExpanded] = useState(false);
+    return (
+      <MenuExpandContext.Provider value={{ expanded }}>
+        Click the accordion to {expanded ? "collapse" : "expand"}
+        <Box maxWidth="200px">
+          <MenuButton.Accordion
+            contentid="content-1"
+            menuButton={{
+              ...args,
+              id: "control-1",
+              "aria-controls": "content-1",
+            }}
+            onClick={() => setExpanded(!expanded)}
+          >
+            <MenuButton label="First item" />
+            <MenuButton label="Second item" />
+          </MenuButton.Accordion>
+        </Box>
+      </MenuExpandContext.Provider>
+    );
+  },
+  args: {
+    label: "Button collapsed",
+    startIcon: HomeIcon,
   },
 };